BRM-336. Provide PARENT_WORKSPACE environment variable for module builds.
In module builds created by the Ivy plugin, WORKSPACE is set to the module workspace.
There are a number of reasons where it might be useful to have access to the workspace
of the job that did the checkout and triggered the module builds. This change sets
a variable named PARENT_WORKSPACE to provide that.